backend/pyproject.toml
[project]
name = "truerisk-backend"
version = "2.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "aiosqlite>=0.21",
    "alembic>=1.14",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "apscheduler>=3.10,<4",
    "xgboost>=2.1",
    "lightgbm>=4.5",
    "scikit-learn>=1.5",
    "torch>=2.4",
    "numpy>=2.0",
    "pandas>=2.2",
    "scipy>=1.14",
    "joblib>=1.4",
    "sse-starlette>=2.0",
    "pywebpush>=2.0",
    "slowapi>=0.1.9",
    "openai>=1.68",
    "twilio>=9.0",
    "pytorch-forecasting>=1.1",
    "pytorch-lightning>=2.4",
    "torch-geometric>=2.6",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.24", "pytest-cov>=6.0", "httpx", "mypy", "ruff"]
training = ["matplotlib>=3.9", "tqdm>=4.66", "tensorboard>=2.18"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
